change small

+++ backend/BookStore.DTO/DTOs/RoleDTO.cs
@@ -12,7 +12,7 @@ namespace BookStore.DTO.DTOs
         public string name { get; set; }
         public string description { get; set; }
         public bool status { get; set; }
-        public List<int> ArrIdPermission { get; set; }
+        public List<long> ArrIdPermission { get; set; }
         public DateTime createdAt { get; set; }
         public DateTime? updatedAt { get; set; }
         public bool isDeleted { get; set; }
